requirements.txt
accelerate==1.10.0
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
einops==0.8.1
fastapi==0.116.1
filelock==3.19.1
fire==0.5.0
fonttools==4.59.1
fsspec==2025.7.0
fugashi==1.5.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
jaconv==0.4.0
jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.9
loguru==0.7.3
manga-ocr==0.1.10
markupsafe==2.1.5
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.3
numpy==2.3.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.1
pillow==9.5.0
pip==25.2
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic-core==2.33.2
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-metric-learning==2.9.0
pytz==2025.2
pyyaml==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
simple-lama-inpainting==0.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.13.3
termcolor==3.1.0
threadpoolctl==3.6.0
timm==1.0.19
tokenizers==0.21.4
tqdm==4.67.1
transformers==4.49.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.182
ultralytics-thop==2.0.16
unidic-lite==1.0.8
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
win32-setctime==1.2.0
